debt - cleanup context key handling in workbench
