+++ src/vs/workbench/browser/parts/notifications/notificationsCenter.ts
@@ -11,7 +11,7 @@ import { INotificationsModel, INotificationChangeEvent, NotificationChangeType, 
 import { IWorkbenchLayoutService, Parts } from 'vs/workbench/services/layout/browser/layoutService';
 import { Emitter } from 'vs/base/common/event';
 import { IContextKeyService } from 'vs/platform/contextkey/common/contextkey';
-import { NotificationsCenterVisibleContext, INotificationsCenterController } from 'vs/workbench/browser/parts/notifications/notificationsCommands';
+import {  INotificationsCenterController } from 'vs/workbench/browser/parts/notifications/notificationsCommands';
 import { NotificationsList } from 'vs/workbench/browser/parts/notifications/notificationsList';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { isAncestor, Dimension } from 'vs/base/browser/dom';
@@ -23,6 +23,7 @@ import { ClearAllNotificationsAction, HideNotificationsCenterAction, Notificatio
 import { IAction } from 'vs/base/common/actions';
 import { IKeybindingService } from 'vs/platform/keybinding/common/keybinding';
 import { assertAllDefined, assertIsDefined } from 'vs/base/common/types';
+import { NotificationsCenterVisibleContext } from 'vs/workbench/common/contextkeys';
 
 export class NotificationsCenter extends Themable implements INotificationsCenterController {
